Fix Dev

+++ TrackingProgressInDevEducationDB/TrackingProgressInDevEducationDB/Tables/Comments.sql
@@ -6,7 +6,7 @@
     [CreatedBy] INT           NOT NULL,
     [TeamId] INT NULL, 
     [CommentFor] INT NOT NULL, 
-    [Estimation] DECIMAL (5,2) NULL, 
+    [Estimation] INT  NULL, 
     CONSTRAINT [PK_COMMENTS] PRIMARY KEY CLUSTERED ([id] ASC),
     CONSTRAINT [Comment_fk0] FOREIGN KEY ([TypeId]) REFERENCES [TrackingProgressInDevEducationDB].[CommentType] ([id]),
     CONSTRAINT [Comment_fk1] FOREIGN KEY ([StudentId]) REFERENCES [TrackingProgressInDevEducationDB].[Students] ([id]),
